+++ src/routes/layoutRoutes.jsx
@@ -1,7 +1,8 @@
 import Dashboard from '../modules/dashboard/pages/dashboard.jsx';
 import { DashboardRoutes } from '../modules/dashboard/router.jsx';
-import UserRoutes from '../modules/product/router.jsx';
+import UserRoutes from '../modules/users/router.jsx';
 import Profile from '../pages-copy/Profile.jsx';
+
 export const LayoutRoutes = [
   {
     title: 'demo',
